Clean up Home + UserCard component (data fetching in Home)
